chore: move landing page template to components
Move LandingPageTemplate out of pages folder into components. Since the
template is used to create pages and is not a page itself it should live
in components.
